compile for cuda/sycl/test/mpi (hip needs trick)

+++ hip/components/cooperative_groups.hip.hpp
@@ -335,7 +335,7 @@ public:
                                                SelectorType selector) const \
     {                                                                       \
         return shuffle_impl(                                                \
-            [this](uint32 v, SelectorType s) {                              \
+            [this](uint16 v, SelectorType s) {                              \
                 return static_cast<const Group*>(this)->_name(v, s);        \
             },                                                              \
             var, selector);                                                 \
@@ -355,12 +355,12 @@ private:
     shuffle_impl(ShuffleOperator intrinsic_shuffle, const ValueType var,
                  SelectorType selector)
     {
-        static_assert(sizeof(ValueType) % sizeof(uint32) == 0,
+        static_assert(sizeof(ValueType) % sizeof(uint16) == 0,
                       "Unable to shuffle sizes which are not 4-byte multiples");
-        constexpr auto value_size = sizeof(ValueType) / sizeof(uint32);
+        constexpr auto value_size = sizeof(ValueType) / sizeof(uint16);
         ValueType result;
-        auto var_array = reinterpret_cast<const uint32*>(&var);
-        auto result_array = reinterpret_cast<uint32*>(&result);
+        auto var_array = reinterpret_cast<const uint16*>(&var);
+        auto result_array = reinterpret_cast<uint16*>(&result);
 #pragma unroll
         for (std::size_t i = 0; i < value_size; ++i) {
             result_array[i] = intrinsic_shuffle(var_array[i], selector);
